+++ backend/src/main/java/com/ziio/backend/controller/NoticeController.java
@@ -1,6 +1,6 @@
 package com.ziio.backend.controller;
 
-import com.ziio.backend.crawling.CrawlerManager;
+import com.ziio.backend.crawler.CrawlerManager;
 import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
